--- app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/home/components/ActionTitle.kt
@@ -1,88 +1,0 @@
-package com.yourssu.soomsil.usaint.screen.home.components
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.outlined.KeyboardArrowRight
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.unit.dp
-import com.yourssu.soomsil.usaint.R
-import com.yourssu.soomsil.usaint.ui.theme.SoomsilUSaintTheme
-
-@Composable
-fun ActionTitle(
-    title: String,
-    subTitle: String,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(
-                horizontal = 16.dp,
-                vertical = 12.dp,
-            ),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Image(
-            modifier = Modifier
-                .size(48.dp)
-                .clip(CircleShape),
-            painter = painterResource(R.drawable.ic_default_profile_image),
-            contentDescription = null,
-        )
-        Column(
-            modifier = Modifier
-                .weight(1f)
-                .padding(horizontal = 12.dp),
-        ) {
-            Text(
-                text = subTitle,
-                style = MaterialTheme.typography.titleSmall.copy(
-                    fontWeight = FontWeight(600),
-                ),
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-            Text(
-                text = title,
-                style = MaterialTheme.typography.titleLarge.copy(
-                    fontWeight = FontWeight(600),
-                ),
-                color = MaterialTheme.colorScheme.onSurface,
-            )
-        }
-        Icon(
-            imageVector = Icons.AutoMirrored.Outlined.KeyboardArrowRight,
-            contentDescription = null,
-            tint = MaterialTheme.colorScheme.onSurface,
-        )
-    }
-}
-
-@PreviewLightDark
-@Composable
-private fun ActionTitlePreview() {
-    SoomsilUSaintTheme {
-        Surface {
-            ActionTitle(
-                title = "전체학기",
-                subTitle = "성적 확인하기",
-            )
-        }
-    }
-}
