--- app/src/main/kotlin/com/yourssu/soomsil/usaint/screen/home/components/GraduationCardItem.kt
@@ -1,41 +1,0 @@
-package com.yourssu.soomsil.usaint.screen.home.components
-
-
-/*@Composable
-private fun GraduationCard(
-    isLoggedIn: Boolean,
-    onGraduationCardClick: () -> Unit,
-) {
-    Surface(
-        rounding = 8.dp,
-        color = YdsTheme.colors.bgNormal,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(bottom = 10.dp),
-        ) {
-            YdsText(
-                text = stringResource(id = R.string.saint_graduation),
-                modifier = Modifier
-                    .padding(
-                        top = 20.dp,
-                        bottom = 4.dp,
-                        start = 16.dp,
-                        end = 16.dp,
-                    ),
-                style = YdsTheme.typography.title3,
-                color = YdsTheme.colors.textPrimary.copy(
-                    alpha = if (isLoggedIn) 1f else DisabledAlpha,
-                ),
-            )
-            ActionTitle(
-                title = stringResource(id = R.string.saint_graduation_title),
-                subTitle = stringResource(id = R.string.saint_graduation_subtitle),
-                onClick = onGraduationCardClick,
-                enable = false,
-                enable = isLoggedIn,
-            )
-        }
-    }
-}*/
